Improve stall detection
Add improved stall detection. It checks that the total remaining is decreasing across checks. 

This can be improved further by adding a configurable minimum speed.

As it is now, it will just ensure that downloads that are progressing won't be canceled due to timeout.

I'm adding this because large albums at a moderate rate can blow the timeout and extending the timeout reduces it's usefulness.